[Kernel][CI] `--jit-monitor-mode error` e2e tests for kernel warmup infra (#50109)
